+++ Giera/SpellProjectile.cpp
@@ -1,5 +1,6 @@
+#include "pch.h"
+
 #include "SpellProjectile.h"
-#include "Drawable.h"
 
 using std::min;
 using std::max;
@@ -60,7 +61,7 @@ void SpellProjectile::move(Time& timeDiff, shared_ptr<Board>& board)
 		Position diff = currPos - prevPos;
 
 		updateAngle(diff);
-		drawables[0]->setAngle(atan2(diff.getY(), diff.getX()) * 57.2957795130823 + 90);
+		drawables[0]->setAngle((float)(atan2(diff.getY(), diff.getX()) * 57.2957795130823 + 90));
 		drawables[0]->setHeightModifier(1 - abs(atan2(diff.grounded().getNorm(), diff.getZ())) * 0.636619772367581);
 		updateHeightModifier(diff);
 		updateDrawables();
